Marketplace: Create DiscoveryPage component
